chore: fix some long running E2E tests (#2508)
* chore: split some e2es

* test: update splits and re-run
