+++ packages/amplify-graphql-api-construct-tests/src/__tests__/deploy-velocity/deploy-velocity-test-core.ts
@@ -5,9 +5,9 @@ import { createNewProjectDir, deleteProjectDir } from 'amplify-category-api-e2e-
 import { AmplifyGraphqlApi } from '@aws-amplify/graphql-api-construct';
 import { initCDKProject, cdkDeploy, cdkDestroy } from '../../commands';
 import { ValidateGraphqlOptions, validateGraphql } from '../../graphql-request';
-import { DURATION_1_HOUR } from './deploy-velocity-constants';
+import { DURATION_90_MINUTES } from './deploy-velocity-constants';
 
-jest.setTimeout(DURATION_1_HOUR);
+jest.setTimeout(DURATION_90_MINUTES);
 
 export type EndpointConfig = Pick<ValidateGraphqlOptions, 'apiEndpoint' | 'apiKey'>;
