+++ scripts/split-e2e-tests.ts
@@ -146,6 +146,8 @@ const RUN_SOLO: (string | RegExp)[] = [
   'src/__tests__/default-ddb-canary.test.ts',
   /src\/__tests__\/group-auth\/.*\.test\.ts/,
   /src\/__tests__\/owner-auth\/.*\.test\.ts/,
+  /src\/__tests__\/relationships\/.*\.test\.ts/,
+  /src\/__tests__\/restricted-field-auth\/.*\.test\.ts/,
 ];
 
 const RUN_IN_ALL_REGIONS = [
